backend/requirements.txt
# Core Flask Dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# PDF Processing (lightweight)
PyMuPDF==1.23.8
python-docx==1.1.0
Pillow==10.2.0

# OCR (basic - Tesseract)
pytesseract==0.3.10

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.31.0

# Azure Services
azure-functions==1.18.0
azure-cosmos==4.5.1

# Database (optional - for MySQL fallback)
mysql-connector-python==8.0.33

# AI Features (Lightweight - ~100 MB total)
# OCR: Tesseract (already included above)
# Text Generation: Groq API
groq==0.4.1
# Embeddings & Vector Search: scikit-learn for TF-IDF
scikit-learn==1.3.2
numpy==1.24.3
scipy==1.11.4
# NLP Processing
nltk==3.8.1
